% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_sql_db.R
\name{append_to_sql_db}
\alias{append_to_sql_db}
\title{Callback function for \code{csv_to_sql_db} that appends new sections to the
SQLite table.}
\usage{
append_to_sql_db(con)
}
\arguments{
\item{con}{A connection to SQLite database.}

\item{x}{Data.frame reading from.}
}
\description{
Callback function for \code{csv_to_sql_db} that appends new sections to the
SQLite table.
}
