% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail_new_dates.R
\name{avail_new_dates}
\alias{avail_new_dates}
\title{Get available data for new format of FCC fixed broadband data}
\usage{
avail_new_dates(user_name, api_key)
}
\arguments{
\item{user_name}{Username for existing FCC account}

\item{api_key}{API key for accessing FCC data. Generated within FCC account.}
}
\value{
tibble of dates available for new FCC data format
}
\description{
Show all dates available for fixed broadband data in the new
format from the FCC
}
\examples{
\dontrun{ avail_new_dates( user_name = "email@location.com", api_key = "longstringofapicharacters")}
}
