% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_sql_db.R
\name{csv_to_sql_db}
\alias{csv_to_sql_db}
\title{FCC file to SQLite Database via chunking}
\usage{
csv_to_sql_db(
  csv_file,
  con,
  pre_process_size = 1000,
  chunk_size = 50000,
  show_progress_bar = TRUE,
  db_colnam
)
}
\arguments{
\item{csv_file}{Name of raw FCC csv file to covert to SQL database. This is
relative  to the working directory unless an absolute file path is provided.}

\item{con}{A DBIConnection object, as returned by \code{dbConnect}.}

\item{pre_process_size}{Number of rows with which to initialize SQL db.}

\item{chunk_size}{Number of rows to include for each chunk}

\item{show_progress_bar}{Display progress bar? (default TRUE)}

\item{db_colname}{Column names for FCC dataset, as returned by
\code{processFCC::get_colname}}
}
\value{
Returns a SQLite database named "table_FCC" inside provided DBI
connection
}
\description{
Reads FCC csv file in chunks while writing each chunk to a SQL
database before reading in the next chunk. DBI connection and FCC column
names from \code{db_colname} function must be provided.
}
\examples{
\dontrun{
# set month and year of FCC data set to use
month = "December"
year = 2020
# download desired dataset
download_FCC(year, month)
# get proper column names
use_colnam <- get_colname(year, month)
# establish database connection
con <- DBI::dbConnect(SQLite(), dbname = "fcc.sqlite")
# filenames are constructed in the following format within \code{download_FCC}
filename <- paste0("FCC_fixed_brdbd_", month, "_", year, ".csv")
csv_to_sql_db(filename, con, db_colnam = use_colnam)
# the SQL database created with \code{csv_to_sql_db} can be accessed in the
# following manner:
table <- tbl(con, "table_fcc")
# close the connection
dbDisconnect(con)
}
}
