% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_FCC.R
\name{download_FCC}
\alias{download_FCC}
\title{Download FCC Fixed Broadband Datasets}
\usage{
download_FCC(year, month)
}
\arguments{
\item{year}{The year of the FCC dataset}

\item{month}{The month of the FCC dataset}
}
\value{
csv (to the working directory) of the desired FCC fixed broadband
dataset
}
\description{
Downloads FCC fixed broadband dataset for the specified year
and month. Years prior to 2020 are available in a zipped format so the
zipped file is downloaded to the working directory, opened, and then the
inside csv is saved to the working directory with a common naming convention
to use later on. The 2020 data sets are only available as a CSV and take
much longer to download (apx. 30+ minutes). Alternatively you can download
the datasets from the FCC website and use the \code{get_colname} function to
get the proper column names for your data downloaded outside of R.
}
\examples{
download_FCC(year = 2019, month = "June")
}
