% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colname.R
\name{get_colname}
\alias{get_colname}
\title{Get Standardized Column Names for FCC Data}
\usage{
get_colname(year, month)
}
\arguments{
\item{year}{The year of the FCC dataset}

\item{month}{The month of the FCC dataset}
}
\value{
vector of the proper column names of the FCC dataset of the
corresponding month and year
}
\description{
Creates vector of proper column names depending on year and
 month of dataset
}
\examples{
new_columns <- get_colname(year = 2019, month = "June")
}
