% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_or_old_FCC.R
\name{new_or_old_FCC}
\alias{new_or_old_FCC}
\title{Is your date of interest in the new or old FCC format?}
\usage{
new_or_old_FCC()
}
\value{
data frame with dates
}
\description{
Return tibble with fixed broadband dates and whether the data is
the new or old format.
}
\examples{
\dontrun{
}
}
