% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup_FCC.R
\name{rollup_FCC}
\alias{rollup_FCC}
\title{Processing FCC data}
\usage{
rollup_FCC(
  con,
  year,
  month,
  state = NULL,
  geogr = "cb",
  tech_exc = c("60", "70"),
  thresh_down = c(25, 25, 100),
  thresh_up = c(3, 5, 100),
  new_file_name = NULL
)
}
\arguments{
\item{con}{A DBIConnection object, as returned by `dbConnect`.}

\item{year}{The year of the FCC data to process.}

\item{month}{The month of the FCC data to process.}

\item{state}{A vector of the state(s) to include in the final data. The
default is NA in order to include all states and territories in the final
data set.}

\item{geogr}{Character representation of Census geography to roll up the
final data set to

  There are 4 options of census geography, listed from smallest to largest:
- cb = Census Block
- cbg = Census Block Group
- ct = Census Tract
- county = County}

\item{tech_exc}{Vector of technology codes to exclude from data when rolling up
if you do not wish to exclude any technology codes input NA or c(NA).}

\item{thresh_down}{Vector of download speeds thresholds. Should be the same
length as `thresh_up` because elements of the vectors will be matched to
count the number of internet providers providing internet at the given
download/upload speed combination within the specified `geogr` region. See
vingette for further explanation of the parameter.}

\item{thresh_up}{Vector of upload speed thresholds. Should be the same
length as `thresh_down` because elements of the vectors will be matched to
count the number of internet providers providing internet at the given
download/upload speed combination within the specified `geogr` region.}

\item{new_file_name}{Name of csv file to output to working directory.}
}
\value{
processed csv file to working directory
}
\description{
Takes connection to SQLite database with FCC file and processes/
rolls up the data based on function inputs.
}
\examples{
\dontrun{
}
}
