% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_map_numprov.R
\name{state_map_numprov}
\alias{state_map_numprov}
\title{Map a column in dataset created by `rollup_FCC`}
\usage{
state_map_numprov(dat, col2map, geogr, yr_dat, state2map, color_pal = "Blues")
}
\arguments{
\item{dat}{Dataset name, created using rollup_FCC, doesn't have to be
subsetted down to `state2map`}

\item{col2map}{name of column in `dat` to map}

\item{geogr}{Character representation of Census geography `dat` is on

  There are 4 options of census geography, listed from smallest to largest:
- cb = Census Block
- cbg = Census Block Group
- ct = Census Tract
- county = County}

\item{yr_dat}{The year of the FCC dataset, `dat`}

\item{state2map}{Two-letter abbreviation of state being mapped}

\item{color_pal}{Colorbrewer Sequential palette option, default is Blues}
}
\value{
map of state with number of providers according to specified column
}
\description{
Creates a geographic map of a state using ggplot2 and tigris
of a numeric column in dataset create using `rollup_FCC`
}
\examples{
state_map_numprov(dat = fcc_dat_2018, col2map = num_prov_25_3,
geogr = "cb", yr_dat = 2018, state2map = "IL", color_pal = "Greens")

}
